+++ server/services/coordinatorService.ts
@@ -1,23 +1,4 @@
-import { UUID } from 'crypto'
-import { CoordinatorApiClient } from '../data'
-
-export interface CreateResponse {
-  sanAssessmentId: UUID
-  sanAssessmentVersion: number
-  sentencePlanId: UUID
-  sentencePlanVersion: number
-}
-
-export interface CreateRequest extends Record<string, unknown> {
-  oasysAssessmentPk: string
-  previousOasysAssessmentPk?: string
-  regionPrisonCode?: number
-  planType: string
-  userDetails: {
-    id: string
-    name: string
-  }
-}
+import CoordinatorApiClient, { CreateRequest } from '../data/coordinatorApiClient'
 
 export default class CoordinatorService {
   constructor(private readonly coordinatorApiClient: CoordinatorApiClient) {}
